Moves BuiltinProgram of the program runtime into its own file.
